src/service/llm/pyproject.toml
[tool.poetry]
name = "llm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.0.48"}
pydantic = "<2"
langchain-openai = "^0.0.8"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
